requirements.txt
streamlit>=1.32.0
opencv-python>=4.8.0
pytesseract>=0.3.10
Pillow>=10.0.0
numpy>=1.24.0
pandas>=2.0.0
supabase>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
